static files zipped and generated in media

+++ WebApp/views.py
@@ -1692,23 +1692,32 @@ class OfflineContentsView(ContentsView):
 import shutil
 import time
 def get_static_files(request):
-    list_of_files = [
-        'static/vendorsx/bootstrap/dist/css/bootstrap.min.css',
-        'static/vendorsx/font-awesome/css/font-awesome.min.css',
-        'static/build/css/theme.min.css',
-        'static/lightbox/css/lightbox.css',
-    ]
+    if not (os.path.exists(os.path.join(settings.MEDIA_URL, 'staticfiles.zip'))):
+        list_of_files = [
+            'static/vendorsx/bootstrap/dist/css/bootstrap.min.css',
+            'static/vendorsx/font-awesome/css/font-awesome.min.css',
+            'static/build/css/theme.min.css',
+            'static/chapterPageBuilder/css/style-content.css',
+            'static/chapterPageBuilder/js/owl-carousel/assets/owl.carousel.css',
+            'static/chapterPageBuilder/js/owl-carousel/owl.carousel.js',
+            'static/images/blankpage.jpg',
+            'static/images/uLMS2019_Loading_SVG.svg',
+            'static/lightbox/css/lightbox.css',
+            'static/lightbox/images/close.png',
+            'static/lightbox/images/loading.gif',
+            'static/lightbox/images/next.png',
+            'static/lightbox/images/prev.png',
+        ]
 
-    path = settings.MEDIA_ROOT
-    for src in list_of_files:
-        dst = os.path.join(path, src)
-        dstfolder = os.path.dirname(dst)
-        if not os.path.exists(dstfolder):
-            os.makedirs(dstfolder)
-        shutil.copy(settings.BASE_DIR+'/WebApp/' +src, dst)
-    zipfile = shutil.make_archive('staticfiles', 'zip', path + '/static' )
-    # time.sleep(2)
-    shutil.make_archive(path + '/staticfiles', 'zip', path + '/static')
+        path = settings.MEDIA_ROOT
+        for src in list_of_files:
+            dst = os.path.join(path, src)
+            dstfolder = os.path.dirname(dst)
+            if not os.path.exists(dstfolder):
+                os.makedirs(dstfolder)
+            shutil.copy(settings.BASE_DIR+'/WebApp/' +src, dst)
+        # time.sleep(2)
+        shutil.make_archive(path + '/staticfiles', 'zip', path + '/static')
 
     return redirect(settings.MEDIA_URL + '/staticfiles.zip')
